aggiunto caso d'uso prenotazioneEsperienza

+++ code/src/main/java/it/unicam/cs/Pagamento.java
@@ -2,11 +2,11 @@ package it.unicam.cs;
 
 public class Pagamento {
 
-    private final int importo;
+    private final float importo;
     private final Turista turista;
     private final Esperienza esperienza;
 
-    Pagamento(int importo, Turista turista, Esperienza esperienza){
+    Pagamento(float importo, Turista turista, Esperienza esperienza){
         this.importo = importo;
         this.turista = turista;
         this.esperienza = esperienza;
@@ -14,7 +14,7 @@ public class Pagamento {
 
     // todo add other methods to complete a payment
 
-    public int getImporto(){
+    public float getImporto(){
        return this.importo;
     }
